+++ utils/constants.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-APP_VERSION = "0.3.008"
+APP_VERSION = "0.3.009"
 
 
 # Player tags